Comments and permissions

+++ Sources/App/Controllers/ReviewController.swift
@@ -7,6 +7,7 @@ final class ReviewController: ResourceRepresentable {
 	}
 	
 	func create(request: Request) throws -> ResponseRepresentable {
+		// Must be logged in
 		let (_, userId) = try request.protected()
 		
 		var review = try request.review(userId: userId)
@@ -20,11 +21,25 @@ final class ReviewController: ResourceRepresentable {
 	}
 	
 	func delete(request: Request, review: Review) throws -> ResponseRepresentable {
+		// Must be logged in
+		// Must own review
+		let (_, userId) = try request.protected()
+		guard review.userId == userId else {
+			throw Abort.custom(status: .forbidden, message: "Can only delete your own review")
+		}
+		
 		try review.delete()
 		return JSON([:])
 	}
 	
 	func update(request: Request, review: Review) throws -> ResponseRepresentable {
+		// Must be logged in
+		// Must own review
+		let (_, userId) = try request.protected()
+		guard review.userId == userId else {
+			throw Abort.custom(status: .forbidden, message: "Can only update your own review")
+		}
+		
 		var review = review
 		try review.patch(node: request.json?.makeNode())
 		try review.save()
